refactor(music-player): create `Analyzer` component
